+++ Server/Code/Routes/tables.routes.ts
@@ -14,6 +14,8 @@ tablesRouter.get("/", (req, res) => {
 
 });
 tablesRouter.post("/", (req, res) => {
+    if(req.auth.role != 'Cashier')
+        return res.status(401).json({ error: "You are not an admin"});
     if (req.body.number == undefined || req.body.seats == undefined || !req.body.isFree || req.body.bill != 0)
         return res.status(400).json({error: "Given Params are not correct"});
     table.newTable(req.body).save().then((table) => {
@@ -25,6 +27,8 @@ tablesRouter.post("/", (req, res) => {
 });
 
 tablesRouter.delete('/:tableid', (req, res) => {
+    if(req.auth.role != 'Cashier')
+        return res.status(401).json({ error: "You are not an admin"});
     if(req.params.tableid == undefined)
         return res.status(404).json({error: "Invalid table id"});
     table.getModel().deleteOne({number: req.params.tableid}).then((table) => {
