+++ Server/Code/Routes/user.routes.ts
@@ -3,6 +3,8 @@ const express = require('express');
 
 export const userRouter = express.Router();
 userRouter.get("/", (req, res) => {
+    if(req.auth.role != 'Cashier')
+        return res.status(401).json({ error: "You are not an admin"});
     user.getModel().find({}).then((users) => {
         return res.status(200).json(users);
     }).catch((err) => {
@@ -10,6 +12,8 @@ userRouter.get("/", (req, res) => {
     });
 });
 userRouter.post("/", (req, res) => {
+    if(req.auth.role != 'Cashier')
+        return res.status(401).json({ error: "You are not an admin"});
     if(req.body.email == undefined || req.body.password == undefined || req.body.username == undefined ||
         (req.body.role != 'Cashier' && req.body.role != 'Waiter' && req.body.role != 'Cook' && req.body.role != 'Bartender'))
         return res.status(400).json({error: "Params given are not correct"});
@@ -25,6 +29,8 @@ userRouter.post("/", (req, res) => {
 });
 
 userRouter.delete("/:email", (req, res) => {
+    if(req.auth.role != 'Cashier')
+        return res.status(401).json({ error: "You are not an admin"});
     const email = req?.params?.email;
     user.getModel().deleteOne({email: email}).then((user) => {
         return res.status(200).json(user);
